refactoring/organizing components
