+++ src/App.css
@@ -24,33 +24,3 @@ body {
 article h1 {
   margin: 1rem 0;
 }
-
-#header {
-  background-color: #181A20
-}
-
-/* For screens with a width between 1024px and 1535px */
-@media (min-width: 0px) and (max-width: 575px) {
-  #header {
-    height: 52px;
-  }
-}
-
-@media (min-width: 576px) and (max-width: 767px) {
-  #header {
-    height: 64px;
-  }
-}
-
-@media (min-width: 768px) and (max-width: 1535px) {
-  #header {
-    height: 66px;
-  }
-}
-
-/* For screens with a width of 1536px and above */
-@media (min-width: 1536px) {
-  #header {
-    height: 85px;
-  }
-}
