+++ src/constants/logos.js
@@ -1,0 +1,2 @@
+export const SMALL_BLUE_LOGO = '/SmallBlueLogoRadicalX.svg';
+export const WHITE_LOGO = '/WhiteLogoRadicalX.svg';
