+++ src/layouts/MainAppLayout/Header/AccountAvatar/styles.js
