+++ src/layouts/MainAppLayout/Header/Logo/styles.js
@@ -1,0 +1,23 @@
+export const logoGridStyle = {
+  alignItems: "center",
+  justifyContent: "center",
+  display: "flex",
+  flexShrink: 0,
+  marginRight: {
+    mobileSmall: "8px",
+    mobile: "24px",
+    tablet: "16px",
+  },
+  img: {
+    width: {
+      mobileSmall: "28px",
+      mobile: "36px",
+      laptop: "166.72px",
+    },
+    height: {
+      mobileSmall: "28px",
+      mobile: "36px",
+      laptop: "21.341px",
+    },
+  },
+};
