+++ src/layouts/MainAppLayout/Header/styles.js
@@ -1,0 +1,68 @@
+export const appBarStyle = {
+  backgroundColor: "#181A20",
+};
+
+export const gridStyle = {
+  display: "flex",
+  justifyContent: "space-between",
+  boxSizing: "border-box",
+  margin: {
+    mobileSmall: "0px 16px 0px 16px",
+    mobile: "0px 24px 0px 24px",
+    tablet: "0px 24px 0px 24px",
+    laptop: "0px 40px 0px 40px",
+    desktop: "0px 42px 0px 40px",
+    desktopLarge: "0px 40px 0px 40px",
+  },
+  height: {
+    mobileSmall: "52px",
+    mobile: "64px",
+    laptop: "66px",
+    desktopLarge: "85px",
+  },
+  backgroundColor: "#181A20",
+};
+
+export const menuBoardStyle = {
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "space-between",
+  gap: {
+    mobileSmall: "auto",
+    mobile: "auto",
+    tablet: "16px",
+    laptop: "16px",
+    desktop: "16px",
+    desktopLarge: "24px",
+  },
+};
+
+export const currenciesStyle = {
+  marginLeft: {
+    mobileSmall: "auto",
+    mobile: "auto",
+    tablet: "0px",
+    laptop: "0px",
+    desktop: "0px",
+    desktopLarge: "0px",
+  },
+  display: "flex",
+  alignItems: "center",
+  alignSelf: {
+    mobileSmall: "center",
+    desktopLarge: "stretch",
+  },
+  gap: {
+    mobileSmall: "4px",
+    mobile: "8px",
+    tablet: "8px",
+    laptop: "8px",
+    desktop: "8px",
+    desktopLarge: "16px",
+  },
+};
+
+export const avatarPhotoStyle = {
+  display: "flex",
+  boxSizing: "border-box",
+};
